+++ police/__resource.lua
@@ -1,9 +1,8 @@
-resource_manifest_version '05cfa83c-a124-4cfa-a768-c24a5811d8f9'
-resource_version '1.4.5'
-resource_versionNum '145'
+resource_manifest_version '44febabe-d386-4d18-afbe-5e627f4af937'
+resource_version '1.4.6'
+resource_versionNum '146'
 resource_Isdev 'no'
-
-dependency 'ghmattimysql'
+resource_fname 'Cops FiveM'
 
 ui_page('client/html/index.html')