+++ police/client/garage.lua
@@ -20,6 +20,7 @@ function load_garage()
 	for k in ipairs (buttons) do
 		buttons [k] = nil
 	end
+
 	for k, data in pairs(vehicles) do
 		if config.useCopWhitelist then
 			if dept == k then
@@ -28,8 +29,10 @@ function load_garage()
 				end
 			end
 		else
-			for k, v in pairs(data) do
-				buttons[#buttons+1] = {name = tostring(v.name), func = "SpawnerVeh", params = tostring(v.model)}
+			if dept == 1 then
+				for k, v in pairs(data) do
+					buttons[#buttons+1] = {name = tostring(v.name), func = "SpawnerVeh", params = tostring(v.model)}
+				end
 			end
 		end
 	end
@@ -53,12 +56,13 @@ function SpawnerVeh(hash)
 	local playerHeading = GetEntityHeading(playerPed)
 	
 	policevehicle = CreateVehicle(car, playerCoords, 90.0, true, false)
+	SetVehicleEngineOn(policevehicle, true, true, true)	
 	SetVehicleMod(policevehicle, 11, 2)
 	SetVehicleMod(policevehicle, 12, 2)
 	SetVehicleMod(policevehicle, 13, 2)
 
 	SetEntityHeading(policevehicle, (playerHeading+160)%360)
-	SetVehicleEnginePowerMultiplier(policevehicle, 35.0)
+	SetVehicleEnginePowerMultiplier(policevehicle, 25.0)
 	SetVehicleOnGroundProperly(policevehicle)
 	SetVehicleHasBeenOwnedByPlayer(policevehicle,true)
 
@@ -67,8 +71,9 @@ function SpawnerVeh(hash)
 	local netid = NetworkGetNetworkIdFromEntity(policevehicle)
 	SetNetworkIdCanMigrate(netid, true)
 	NetworkRegisterEntityAsNetworked(VehToNet(policevehicle))
+
 	TaskWarpPedIntoVehicle(playerPed, policevehicle, -1)
-	SetEntityInvincible(policevehicle, false)
+	SetEntityInvincible(policevehicle, false)	
 end
 
 function OpenGarage()